Move helper macros to CustomTestMacros
