+++ tests/Feature/Reporting/UnacceptedAssetReportTest.php
@@ -3,44 +3,10 @@
 namespace Tests\Feature\Reporting;
 
 use App\Models\User;
-use Illuminate\Testing\TestResponse;
-use League\Csv\Reader;
-use PHPUnit\Framework\Assert;
 use Tests\TestCase;
 
 class UnacceptedAssetReportTest extends TestCase
 {
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        TestResponse::macro(
-            'assertSeeTextInStreamedResponse',
-            function (string $needle) {
-                Assert::assertTrue(
-                    collect(Reader::createFromString($this->streamedContent())->getRecords())
-                        ->pluck(0)
-                        ->contains($needle)
-                );
-
-                return $this;
-            }
-        );
-
-        TestResponse::macro(
-            'assertDontSeeTextInStreamedResponse',
-            function (string $needle) {
-                Assert::assertFalse(
-                    collect(Reader::createFromString($this->streamedContent())->getRecords())
-                        ->pluck(0)
-                        ->contains($needle)
-                );
-
-                return $this;
-            }
-        );
-    }
-
     public function test_permission_required_to_view_unaccepted_asset_report()
     {
         $this->actingAs(User::factory()->create())
